reduce array copies in patch tracking
When Harmony tracks patches for a method, it copies the patch list twice for each added patch (copy array into a list, add one entry, copy it back into an array, repeat). This commit reduces that to two copies regardless of the number of patches being added, and reduces duplicate code.

PatchInfo isn't technically part of the API, but this keeps backwards compatibility just in case per discussion since the class is public.