Use `is` over `==`/`!=` for null/false checks
Besides being more idiomatic in modern C#, this marginally improves performance for types that implement equality operators (such as Type, FieldInfo, etc.).
As for the `!= null` replacement, the `!(... is null)` syntax is too cumbersome, so I went with the kosher yet slightly confusing `is object` pattern.
When C# 9 is released, this can be changed to `is not null`.