HarmonySharedState GetPatchedMethods returns array (method signature unchanged)
Dictionary.Keys is not thread-safe and locking the dictionary won't work either since it's a lazy enumerator
Only solution is to "resolve" it into an array or list, or use ConcurrentDictionary (which isn't available in net35).
This should have negligable performance impact since users that call this method typically iterate over the whole collection.