Fix issue 414: add a sacrificial JIT trampoline after short methods
The previous attempt to fix the short method issue was allocating memory
from a dynamic code manager. However, the detour was overwriting memory
inside the global code manager.

This fix calls the dynamically generated padding method, which forces
Mono to allocate a JIT trampoline for the padding method. Even though
the padding function is stored in a dynamic code manager, the JIT
trampoline is allocated from the global code manager. Shortly
afterwards, patching the short function will overwrite the padding
method's JIT trampoline, instead of the next important thing that is
stored in the global code manager.