Ensure build works when .NET Core 3.1 or 5.0 is not installed
ILRepack usually resolves framework assemblies from the dotnet SDK directories if the corresponding .NET Core version is installed. If it's not installed, let's download Microsoft.NetCore.App.Ref NuGet package (equivalent of Microsoft.NETFramework.ReferenceAssemblies but for Core) of the specified version and pass the path to ILRepack so it can find the .dlls there.

Confirmed this fixes the build.